Preparation for generic user views
Added a getSuffixValues method that is meant to allow for having a
non-public field as key_name.
Also added support for a key-name field being embeded in the form so
that if the key-fields change the old key_name is not lost.

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%401000