Added agreed_to_tos_on field in user and role model.
Also the user and user_self view no properly handle this new field.

The ToS will be shown on-screen (not in a nice way) only if it is in effect and a user has not signed it yet.

Patch by: Lennard de Rijk
Reviewed by: to-be-reviewed

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%401683