Cleanup in org_app and replace ReferenceProperty with LinkProperty
Using LinkProperty allows the org to host their idea's list on
Melange through the Document system, and then specify the URL of that
document. This also makes it possible for the orgs to specify a
document that is not hosted on Melange.

Patch by: Sverre Rabbelier

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%401847