Added apply as a mentor view which lists all orgs for a given program.
The check if someone can actually request to become a mentor is done after the user has selected an organization from this view. This is not bad because a student would never see the link to this page in the program menu.

Patch by: Lennard de Rijk
Reviewed by: to-be-reviewed

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%402035