Added two menu items to the program menu.
The first addition is a link to redirect to the Host's view for showing duplicates.

The second addition is a link for Students to an overview of all their projects in the current program.

Patch by: Lennard de Rijk
Reviewed by: to-be-reviewed

--HG--
extra : convert_revision : svn%3A32761e7d-7263-4528-b7be-7235b26367ec/trunk%402901