Use offset_linkid instead of offset to scan >1000 entities.
this is a first-cut.  It works in all the ways I could make earlier
versions fail.  It passes link_id as URL parameters. It also has a new
class LinkCreator which makes the main body of getListContents even easier
to write.

I wasn't sure if link_id's could have non alphanumeric characters; if so, they
need to be URL encoded/decoded.


I also need to go and remove any mention of raw offsets now, because we don't
use them.

I believe I've talked about this approach with a few of you and it sounded
reasonable.  Feel free to roll-back/fix/amend/comment-for-me-to-fix. This is
my first big-logic-change to Melange.

Patch by: Dan Bentley

--HG--
branch : gae-fetch-limitation-fix