Added getSurveyResponseFromPost method to Surveys Helper.
This method will be used to read the POST dictionary from a request. And will transform it into the fields that need to be set in a SurveyRecord for the questions specified by the given Survey.

A TODO has been added to handle the naming/handling of the comment fields.