Added list of Surveys to the Project's manage page.
This list will show wether or not a Survey has been taken and will also feature links to (re)take a Survey and in the near future to see the actual SurveyRecord.

Note that the take Survey link is working however Org Admins will receive an access denied message when following this link.