Argument to_json added to json function in Base view.
Its value (which should be a boolean) determines if the data should be converted to a JSON object. Therefore if someone has already a JSON string and wants to return it in a HTTP response, he or she may set to_json=False and the object will not be converted.