Adds buttons to notifications list.
The buttons are Mark as read, Mark as unread and Delete buttons. Also adds
checkbox to each of the notification in the list. So once the user selects
a number of notifications using checkbox and clicks one of the options
the action performed as per the button clicked will be performed.

Fixes issue 450.