Add explicit GPLv2 and LGPLv2 option to the licenses
Currently the list of open source licenses in models/licenses.py, which is
pulled from [0], lists "(L)GPL" and "(L)GPLv3", but does not include the
"(L)GPLv2", however, on [0] the "(L)GPL" one is linked to the "(L)GPLv2"
page, so it is not unfair to assume that the "(L)GPLv2" was intended.

As such, add an explicit "(L)GPLv2" license.

[0] http://www.opensource.org/licenses/alphabetical