Add minimized versions of all js files
Running the closure compiler takes a long time, and is only needed
when a js file has changed. This also removes the need to break the
symlink in the build directory for "soc".

As such, break the "do not put generated files in your vcs" and add
the minimized version of all our js files in a *.min directory.