Add tests to the role logic
Set up the environment for role logic tests.

Added a test case to test creating a role for a user.

Added a test case to test updating the information of role for user.

Added a test case to test deleting a role of a user.

Added test cases to test getSuggestedInitialProperties() method.