Users must be able to apply as mentors for multiple Organizations.
Currently with the conditions the sidebar link used to exist for users
who have not applied to become mentors even for a single Organization. The
moment they become mentors for one Organization the link would not show up.
Now apply to become a mentor shows until the program end even if the user
is already a mentor for some organization.