Always send new notification email
Since we now include the fulltext of the notification in the
notification email it is more likely that users do not read the
message online. This means that it is likely that users end up not
seeing that they have a new notification.

Fix this by always sending out notification emails.