Enclose all table cell texts in links.
This overrides the "click on row" using Javascript, so
now the user can decide whether to open the link in the
current tab or in a new one (or a new window).
Provides stripping of link tags when exporting to CSV.
Known bug: the link doesn't cover the whole table cell.

Fixes Issue 740. (hopefully)