Time to complete has been split into days and hours field on the task create/edit/suggest forms.
However this field on the datastore side is unchanged. So we
have to normalize the form data before storing it in the
datastore and do the reverse procedure while constructing the
form.

Also, now there are initial values of 0 for time to complete
days field and 1 for time to complete hours fields.

Fixed issue 1009.