Send form_error get variable in URL when form has errors during file uploads and delete blobs.
When the form involved has errors(validation) the data will not be
saved. However it is not possible to achieve this in the usual way
because of the blobstore API restriction that it always expects an
HTTP Redirect and not a response. So we are tying to circumvent this
problem here by setting a GET variable form_error in the URL and send
a redirect to that URL which is basically the task public get page.

This is prone to problems like say, a user for fun may enter
form_error=1 in the URL to fake this error on the form with
file uploads.

This patch also deletes the blobs when the form is not saved.

Fix issue 991.