Giving additional time when Task is in NeedsReview and deadline passed no more gives 500.
Currently when the task is in NeedsReview state and the deadline passes
and then the mentor asks for more work and tries to give an additional
amount of time, there is a 500 because the deadline for the Task is
already unset. This patch fixes it by starting the deadline counter from
the current moment.