Change the student apply URL used in the program view to use the redirect function.
This change is done to be consistent with the way all the URLs are
fed into the views. In almost every other place in Melange, we use
the redirect function from the appropriate redirect helper module.
So student apply URL here should be in consistent with other places
of usage too.