Additional Model properties for connect with us links for the GSoC Program.
The properties are added as LinkProperty or EmailProperty and all of them
have required=False, meaning they are optional. When there is no value
entered for one of these specific properties on the Edit Program Profile
Form, that corresponding icon won't be displayed on the Program home page.

Also, we are using url cleaners for LinkProperty and empty field cleaner
for Email Property.

Important thing to note here is that we use EmailProperty for IRC since
we follow the convention of irc://<channel>@<server>