a few minor bugs in validate function in Profile view fixed.
Because of the construction:
if self.data.role and student_info_form.is_valid():

the student info form was not always validated.

Also, when profile was created, it was not populated to self.data.role, so it is replaced with profile.

--HG--
extra : rebase_source : 6284939363c7935a58a708ce3e5aa35e663fa0ae