ListConfiguration is now purely a configuration.
The ListConfigurationResponse will know how to render the dictionary containing the configuration.

The ListContentResponse knows how to render each row of data, with the exception of buttons and row actions which need to be further defined. This class can be used by a class that holds a query and Logic instance to push the next batch of entities through addRow.