Add sort orders for all the lists except a few.
The default sort orders are not added for student proposals, projects
lists and also for survey lists intentionally since students may expect
to see them listed on the order in which they submitted it.

But currently we use the key to sort by default, this may not be the
expected behavior too, but I will leave them as is until we add the
functionality to sort them on the date of submission.

--HG--
extra : rebase_source : 5515b47c72133f449f81bccfc1c4b9a02233b1e7